+++ Gyrfalcon/DesktopClient.ManualTest/FirstFly.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.IO;
 using DesktopClient.ProcessMonitor;
 using DesktopClient.SystemServices;
+using DataStore;
 
 namespace DesktopClient.ManualTest
 {
@@ -61,9 +62,9 @@ namespace DesktopClient.ManualTest
 
 			for (int i = 0; i < csResult.GetLength(0); i++)
 			{
-				aggregatedDataList[csResult[i, 0]].Frequency = csResult[i, 1];
+				aggregatedDataList[(int)csResult[i, 0]].Frequency = csResult[i, 1];
 
-				finalDataList.Add(aggregatedDataList[csResult[i, 0]]);
+				finalDataList.Add(aggregatedDataList[(int)csResult[i, 0]]);
 			}
 
 			using (StreamWriter sw = File.AppendText("ClassifiedAnalysis.txt"))
